chore(e2e) rename e2e-test-page -> page
