+++ packages/core/scripts/e2e/page.js
